% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRS.R
\name{AccrualCalculation}
\alias{AccrualCalculation}
\title{A function that calculates the accrual amounts for each leg}
\usage{
AccrualCalculation(swap.dates, leg.type, swap, direction, floating.history)
}
\arguments{
\item{swap.dates}{A list of lists with the main cashflow information for
both the legs}

\item{leg.type}{A char that describes whether the leg is payer or
receiver}

\item{swap}{A list with the swap's charachteristics}

\item{direction}{A parameter that gets a value of 1 for receiver swaps and -1
for payer swaps}

\item{floating.history}{A table downloaded automatically from Quandl with the
historical floating rates}
}
\value{
The accrual amount
}
\description{
This function calculates the accrual for each leg by also downloading the
floating leg paramater using the Quandl package and datafeed. Currently the
function considers only EUR contracts on 6 months floating leg
}
