% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRS.R
\name{CalculateAccrual}
\alias{CalculateAccrual}
\title{A function that calculates the accrual amounts for each leg}
\usage{
CalculateAccrual(swap.dates, leg.type, swap, direction)
}
\arguments{
\item{swap.dates}{A list of lists with the main cashflow information for
both the legs}

\item{leg.type}{A charachter that describes whether the leg is payer or
receiver}

\item{swap}{A list with the swap's carachteristics}

\item{direction}{A parameter that gets a value of 1 for receiver swaps and -1
for payer swaps}
}
\value{
The accrual amount
}
\description{
This function calculates the accrual for each leg by also downloading the
floating leg paramater using the Quandl package and datafeed. Currently the
function considers only EUR contracts on 6 months floating leg
}
