% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRS.R
\name{CashflowCalculation}
\alias{CashflowCalculation}
\title{A function that calculates the main set of dates for a given leg of the
contract}
\usage{
CashflowCalculation(today, start.date, maturity.date, type, time.unit, dcc,
  calendar)
}
\arguments{
\item{today}{The Date at which the analysis is being carried out}

\item{start.date}{The settlement date of the contract}

\item{maturity.date}{The maturity date of the contract}

\item{type}{Type of leg (ie. floating or fixed)}

\item{time.unit}{Number of months for the frequency of the leg (ie. monthly
would have a time.unit of 1, quarterly of 3, semiannual of 6 and annual of 12)}

\item{dcc}{Day Count Convention as per the RQuantLib doc}

\item{calendar}{Character with the holiday's calendar as per the RQuantLib doc}
}
\value{
A list which contains 1) future cashflows 2) starting date
of the current accrual period 3) the fixing date for the variable rate
}
\description{
This function calculates the following set of dates (then converted into year
fractions from the valuation date if needed): 1) future cashflows 2) starting
date of the current accrual period 3) the fixing date for the variable rate
}
