% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRS_Optimised.R
\name{CashFlowPricing}
\alias{CashFlowPricing}
\title{Main function that manages the pricing of one interest rate swap}
\usage{
CashFlowPricing(today, swap, df.table)
}
\arguments{
\item{today}{The Date at which the analysis is being carried out}

\item{swap}{A list with the swap's carachteristics}

\item{df.table}{A tibble with the discount factor curve information}
}
\value{
A list with all the main pricing information for the contract
}
\description{
This function connects the cashflow generation to the swap pricing
}
