% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.table}
\alias{df.table}
\title{A discount factor term structure}
\format{A tibble with 26 rows and 2 columns:
\describe{
   \item{t2m}{a day fraction representing the time to maturity of the
   instruments used for constructing the discount factor curve}
   \item{df}{the level of the discount factor for that time to maturity}
}}
\source{
Bloomberg SWPM page as at the 15th of April 2019
}
\usage{
df.table
}
\description{
A table containing the discount factor term structure
}
\keyword{datasets}
