% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_formatting.R
\name{GetStandard}
\alias{GetStandard}
\title{Helper function to format standard swaps}
\usage{
GetStandard(swap.tabular, leg, variable)
}
\arguments{
\item{swap.tabular}{a list which includes the user defined characteristics
of a swap contract}

\item{leg}{specifies whether the leg is payer or receiver}

\item{variable}{the contract feature that is being converted}
}
\value{
a standard feature for the contract
}
\description{
This function calls other helper functions to format a swap list from the
tabular version to the target one
}
