% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_formatting.R
\name{GetStandardCalendar}
\alias{GetStandardCalendar}
\title{Helper function to format standard swaps}
\usage{
GetStandardCalendar(swap.tabular)
}
\arguments{
\item{swap.tabular}{a list which includes the user defined characteristics
of a swap contract}
}
\value{
the standard calendar for the currency
}
\description{
This function calls other helper functions to format a swap list from the
tabular version to the target one
}
