% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_formatting.R
\name{GetStandardList}
\alias{GetStandardList}
\title{Helper function to format standard swaps}
\usage{
GetStandardList(swap.tabular, variable)
}
\arguments{
\item{swap.tabular}{a list which includes the user defined characteristics
of a swap contract}

\item{variable}{the contract feature that is being converted}
}
\value{
a flist of standard characteristics for both the legs
}
\description{
This function calls other helper functions to format a swap list from the
tabular version to the target one
}
