% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRS.R
\name{OLDParSwapRate}
\alias{OLDParSwapRate}
\title{Calculates the Par Swap Rate and Annuity}
\usage{
OLDParSwapRate(swap.cf)
}
\arguments{
\item{swap.cf}{A 2 column swap cashflow tibble with coupon dates (in terms of
year fractions) and discount factors}
}
\value{
A list containing the par swap rate and the annuity
}
\description{
Calculates the Par Swap Rate and Annuity using  the old "one curve"
construction methodology
}
