% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRS.R
\name{OLDParSwapRateCalculation}
\alias{OLDParSwapRateCalculation}
\title{A function that prepares the data for the swap rate and annuity calculation}
\usage{
OLDParSwapRateCalculation(swap.dates, swap, df.table)
}
\arguments{
\item{swap.dates}{A list of lists with the main cashflow information for
both the legs}

\item{swap}{A list with the swap's carachteristics}

\item{df.table}{A tibble with the discount factor curve information}
}
\value{
A list containing the par swap rate and the annuity
}
\description{
This function in particular selects the leg that is paying fixed as the old
calculation methodology considers the discount factor curve and the forward
curve to be the same. This makes the floating leg always pricing at par. It
then interpolates the discount factor over the cashflow dates using the log
linear interpolation methodology.
}
