% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRS.R
\name{SwapCashFlowCalculation}
\alias{SwapCashFlowCalculation}
\title{Main function that manages the pricing of one interest rate swap}
\usage{
SwapCashFlowCalculation(today, swap)
}
\arguments{
\item{today}{The Date at which the analysis is being carried out}

\item{swap}{A list with the swap's carachteristics}
}
\value{
A list with all the main pricing information for the contract
}
\description{
This function connects the cashflow generation to the swap pricing
}
