% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_formatting.R
\name{SwapPortfolioFormatting}
\alias{SwapPortfolioFormatting}
\title{Formatting the swap table}
\usage{
SwapPortfolioFormatting(swap.tabular)
}
\arguments{
\item{swap.tabular}{a list which includes the user defined characteristics
of a swap contract}
}
\value{
a formatted version of the contract that is readable by the functions
in the toolbox
}
\description{
This function converts a tabular input into the relevant lists
}
