% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USD.curves}
\alias{USD.curves}
\title{USD interest rate curve set}
\format{A list with two elements :
\describe{
   \item{currency}{with the ISO code of the currency (USD)}
   \item{discount}{a tibble containing the following two columns:}
   \itemize{
         \item t2m: a day fraction representing the time to maturity of the
   instruments used for constructing the discount factor curve
         \item df: the level of the discount factor for that time to maturity
   }
}}
\source{
Bloomberg SWPM page as at the 15th of April 2019
}
\usage{
USD.curves
}
\description{
A list containing the currency of the curve and the discount factor term structure
}
\keyword{datasets}
