% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRS.R
\name{SwapPricing}
\alias{SwapPricing}
\title{A function that calculates the main characteristics of the contract}
\usage{
SwapPricing(swap.dates, swap, df.table, floating.history)
}
\arguments{
\item{swap.dates}{A list of lists with the main cashflow information for
both the legs}

\item{swap}{A list that contains the contract information}

\item{df.table}{A tibble with the discount factor curve information}

\item{floating.history}{A df with the historical data downloaded from Quandl}
}
\value{
A list with all the main pricing information for the contract
}
\description{
This function calculates the market value, accruals and PV01. At the moment
it only prices swaps using a "one curve" methodology which is the old way of
pricing this type of contracts
}
